requirements.txt
asgiref==3.2.10
Django==3.1.2
django-jazzmin==2.2.11
django-tinymce==3.1.0
djangorestframework==3.12.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
